refactor: move tokens to enums
