+++ src/enums.rs
@@ -6,6 +6,24 @@ pub enum FileType {
     Toml,
 }
 
+#[derive(Debug, Clone, PartialEq)]
+pub enum Token {
+    LeftBrace,    // {
+    RightBrace,   // }
+    LeftBracket,  // [
+    RightBracket, // ]
+    Colon,        // :
+    Comma,        // ,
+    Equals,       // =
+    Dot,          // .
+    String(String),
+    Number(f64),
+    Boolean(bool),
+    DateTime(String),
+    Null,
+    EOF,
+}
+
 impl fmt::Display for FileType {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
