--- src/parser/token.rs
@@ -1,17 +1,0 @@
-#[derive(Debug, Clone, PartialEq)]
-pub enum Token {
-    LeftBrace,    // {
-    RightBrace,   // }
-    LeftBracket,  // [
-    RightBracket, // ]
-    Colon,        // :
-    Comma,        // ,
-    Equals,       // =
-    Dot,          // .
-    String(String),
-    Number(f64),
-    Boolean(bool),
-    DateTime(String),
-    Null,
-    EOF,
-}
